[ADF-3749] Refactored cloud demo components
