+++ demo-shell/src/app/components/cloud/cloud-breadcrumb-component.html
@@ -1,7 +1,7 @@
-<mat-toolbar>
+<adf-toolbar>
 <div fxLayout="column" fxLayoutAlign="center" >
   <div fxLayout="row">
     <div class="appCrumb">{{applicationName + ' >'}} </div>  <div class="filterCrumb"> {{filterName}}</div>
   </div>
 </div>
-</mat-toolbar>
+</adf-toolbar>
