+++ demo-shell/src/app/components/cloud/cloud-breadcrumb-component.scss
@@ -1,6 +1,7 @@
 .appCrumb {
   opacity: .5;
 }
+
 .filterCrumb {
   opacity: .8;
   margin-left: 5px;
