+++ demo-shell/src/app/components/cloud/task-list-cloud-demo/task-list-cloud-demo.component.scss
